Uncover gene modules and their regulatory programs from single-cell data

Use the scRegClust algorithm to determine gene modules and their regulatory programs from single-cell data.

Usage

scregclust(
  expression,
  genesymbols,
  is_regulator,
  penalization,
  n_cl,
  target_cluster_start = NULL,
  sample_assignment = NULL,
  center = TRUE,
  split1_proportion = 0.5,
  total_proportion = 1,
  split_indices = NULL,
  prior_indicator = NULL,
  prior_genesymbols = NULL,
  prior_baseline = 1e-06,
  prior_weight = 0.5,
  min_cluster_size = 0L,
  allocate_per_obs = TRUE,
  noise_threshold = 0.025,
  n_cycles = 50L,
  use_kmeanspp_init = TRUE,
  n_init_clusterings = 50L,
  max_optim_iter = 10000L,
  tol_coop_rel = 1e-08,
  tol_coop_abs = 1e-12,
  tol_nnls = 1e-04,
  compute_predictive_r2 = TRUE,
  compute_silhouette = FALSE,
  verbose = TRUE
)

Arguments

expression

p x n matrix of pre-processed single cell expression data with p rows of genes and n columns of cells.

genesymbols

A vector of gene names corresponding to rows of expression. Has to be of length p.

is_regulator

An indicator vector, telling which rows in expression are candidate regulators. Has to be of length p.

penalization

Sparsity penalty controlling the amount of regulators used for each cluster. Either a single positive number or a vector of positive numbers.

n_cl

Requested number of clusters (integer). If this is provided without specifying target_cluster_start, then an initial clustering is performed on the cross-correlation matrix of targets and genes on the first dataset after data splitting.

target_cluster_start

The start cluster assignment for the rows in expression that correspond to target genes, i.e. those for which is_regulator == 0L. If this is not specified, then see n_cl regarding cluster initialization. If this is provided then use_kmeanspp_init and n_init_clusterings are ignored.

sample_assignment

A vector of sample assignment for each cell, can be used to perform the data splitting with stratification. Has to be of length n. No stratification if NULL is supplied.

center

Whether or not genes should be centered within each subgroup defined in sample_assignment.

split1_proportion

The proportion to use for the first dataset during data splitting. The proportion for the second dataset is 1 - split1_proportion. If stratification with sample_assignment is used, then the proportion of each strata is controlled.

total_proportion

Can be used to only use a proportion of the supplied observations. The proportion of the first dataset during data splitting in relation to the full dataset will be total_proportion * split1_proportion.

split_indices

Can be used to provide an explicit data split. If this is supplied then split1_proportion, and total_proportion are ignored. Note that if sample_assigment is provided and center == TRUE, then subgroup centering will be performed as in the case of random splitting. A vector of length n containing entries 1 for cells in the first data split, 2 for cells in the second data split and NA for cells that should be excluded from the computations.

prior_indicator

An indicator matrix (sparse or dense) of size q x q that indicates whether there is a known functional relationship between two genes. Ideally, this is supplied as a sparse matrix (sparseMatrix in the Matrix package). If not, then the matrix is converted to one.

prior_genesymbols

A vector of gene names of length q corresponding to the rows/columns in prior_indicator. Does not have to be the same as genesymbols, but only useful if there is overlap.

prior_baseline

A positive baseline for the network prior. The larger this parameter is, the less impact the network prior will have.

prior_weight

A number between 0 and 1 indicating the strength of the prior in relation to the data. 0 ignores the prior and makes the algorithm completely data-driven. 1 uses only the prior during cluster allocation.

min_cluster_size

Minimum required size of target genes in a cluster. Smaller clusters are emptied.

allocate_per_obs

Whether cluster allocation should be performed for each observation in the second data split separately. If FALSE, clusters are allocated on the aggregate sum of squares across all observations in the second data split

noise_threshold

Threshold for the best \(R^2\) of a target gene before it gets identified as noise.

n_cycles

Number of maximum algorithmic cycles.

use_kmeanspp_init

Use kmeans++ for cluster initialization if target_cluster_start is a single integer; otherwise use kmeans with random initial cluster centers

n_init_clusterings

Number of kmeans(++) initialisation runs.

max_optim_iter

Maximum number of iterations during optimization in the coop-Lasso and NNLS steps.

tol_coop_rel

Relative convergence tolerance during optimization in the coop-Lasso step.

tol_coop_abs

Absolute convergence tolerance during optimization in the coop-Lasso step.

tol_nnls

Convergence tolerance during optimization in the NNLS step.

compute_predictive_r2

Whether to compute predictive \(R^2\) per cluster and regulator importance

compute_silhouette

Whether to compute silhouette scores for each target gene.

verbose

Whether to print progress.

Value

A list with S3 class scregclust containing

penalization

The supplied penalization parameters

results

A list of result lists (each with S3 class scregclust_result), one for each supplied penalization parameter. See below.

target_cluster_start

Initial clustering for target genes, i.e. those with is_regulator == 0

split_indices

either verbatim the vector given as input or a vector encoding the splits as NA = not included, 1 = split 1 or 2 = split 2. Allows reproduciblity of data splits.

For each supplied penalization parameter, results contains a list with

the current penalization parameter,

the supplied genesymbols after filtering (as used during fitting),

the supplied is_regulator vector after filtering (as used during fitting),

the number of fitted clusters n_cl,

whether the current run converged to a single configuration (as a boolean),

as well as an output object containing the numeric results for each final configuration.

It is possible that the algorithm ends in a cycle instead of a unique final configuration. Therefore, output is a list with each element itself being a list with the following contents:

reg_table

a regulator table, a matrix of weights for each regulator and cluster

cluster

the cluster assignments for all genes with regulators marked as NA

cluster_all

the cluster assignments for all genes with regulators marked as NA. Genes that are in the rag-bag cluster (marked with -1 in cluster) are assigned to the existing cluster in which it has the largest \(R^2\), even if it is below noise_threshold.

r2

matrix of predictive \(R^2\) value for each target gene and cluster

best_r2

vector of best predictive \(R^2\) for each gene (regulators marked with NA)

best_r2_idx

cluster index coreesponding to best predictive \(R^2\) for each gene (regulators marked with NA)

r2_cluster

a vector of predictive \(R^2\) values for each cluster (included if compute_predictive_r2 == TRUE)

importance

a matrix of importance values for each regulator (rows) and cluster (columns) (included if compute_predictive_r2 == TRUE)

r2_cross_cluster_per_target

a matrix of cross cluster \(R^2\) values for each target gene (rows) and each cluster (columns) (included if compute_silhouette == TRUE)

silhouette

a vector of silhouette scores for each target gene (included if compute_silhouette == TRUE)

models

regulator selection for each cluster as a matrix with regulators in rows and clusters in columns

signs

regulator signs for each cluster as a matrix with regulators in rows and clusters in columns

weights

average regulator coefficient for each cluster

coeffs

list of regulator coefficient matrices for each cluster for all target genes as re-estimated in the NNLS step

sigmas

matrix of residual variances, one per target gene in each cluster; derived from the residuals in NNLS step